CASMINST-2940: check bootparameters instead of bootscript for no-wipe
The call to retrieve the bootscript has the potential to return
a chained request.  This always happens for ncn-m001, such that
the script has a different procedure it uses for ncn-m001. This
procedure would work for ncn-s001 as well in this case.

Checking the bootparameters instead of the bootscript allows us
to use the same check for all NCNs.